event-consumer/requirements.txt
kafka-python==2.0.2
redis==5.0.1

# Testing
pytest==7.4.3
pytest-mock==3.12.0
